CHORE: refactor nav
So much for the stuff we don't even use. There was a system for multi-level sub menu's?
That was way too overkill and once you have that, you likely have a bad site design. It would make navigation be quite bad on the site.
Left a comment in case this gets revisited in the future.

Also updated the buttons a tiny bit